985717: PR-4: empty cron job exception fix

+++ ext/marketplacecommerceservices/src/com/tisl/mpl/marketplacecommerceservices/jobs/SendRefundSmsCronJob.java
@@ -43,75 +43,45 @@ public class SendRefundSmsCronJob extends AbstractJobPerformable<CronJobModel>
 		LOG.debug("Inside perform method of SendRefundSmsCronJob class...");
 		try
 		{
-			final String query = refundSmsDao.getAllTransactionsForSms();
-			final List<BulkSmsPerBatch> result = refundSmsDao.searchResultsForRefund(query);
-			final int rowCount = result.size();
-			final int batch = configurationService.getConfiguration().getInt("bulksms.perbatch.sms");
-			int dividedValue = 0;
-			dividedValue = rowCount / batch;
-			int remStart = 0;
-			int remEnd = 0;
-			int remainder = 0;
-			//For removing remainder loop
-			if (dividedValue == 0 && rowCount != 0)
+			final boolean onOffSwitch = configurationService.getConfiguration().getBoolean("refund.sms.switch");
+			if (onOffSwitch)
 			{
-				remainder = rowCount;
-				remStart = 0;
-				remEnd = rowCount;
-			}
 
-			if (dividedValue * batch < rowCount)
-			{
-				remainder = rowCount - (dividedValue * batch);
-			}
-			final StringBuilder deleteDynamicQuery = new StringBuilder();
-			int a = 0;
-			int b = batch;
-			List<BulkSmsPerBatch> result1 = null;
-			boolean response = false;
-			LOG.debug("========STEP:1==============");
-			//parent loop
-			for (int j = 1; j <= dividedValue; j++)
-			{
-				result1 = null;
-				response = false;
-				//for (int i = 1; i <= b; i++)
-				//{
-				result1 = result.subList(a, b);
-				response = sendSmsService.sendBulkSms(result1);
-				//}
-				///delete dynamic query
-				if (response)
+				final String query = refundSmsDao.getAllTransactionsForSms();
+				final List<BulkSmsPerBatch> result = refundSmsDao.searchResultsForRefund(query);
+				final int rowCount = result.size();
+				final int batch = configurationService.getConfiguration().getInt("bulksms.perbatch.sms");
+				int dividedValue = 0;
+				dividedValue = rowCount / batch;
+				int remStart = 0;
+				int remEnd = 0;
+				int remainder = 0;
+				//For removing remainder loop
+				if (dividedValue == 0 && rowCount != 0)
 				{
-					for (final BulkSmsPerBatch obj : result1)
-					{
-						deleteDynamicQuery.append("'");
-						deleteDynamicQuery.append(obj.getTxnId());
-						deleteDynamicQuery.append("'");
-						deleteDynamicQuery.append(",");
-					}
-				}
-				a += batch;
-				b += batch;
-				if (b > rowCount)
-				{
-					remStart = rowCount - a;
+					remainder = rowCount;
+					remStart = 0;
 					remEnd = rowCount;
-					break;
 				}
-			}
-			LOG.debug("========STEP:2==============");
-			//Remainder loop
-			if (remainder != 0)
-			{
-				for (int j = 1; j <= 1; j++)
+
+				if (dividedValue * batch < rowCount)
+				{
+					remainder = rowCount - (dividedValue * batch);
+				}
+				final StringBuilder deleteDynamicQuery = new StringBuilder();
+				int a = 0;
+				int b = batch;
+				List<BulkSmsPerBatch> result1 = null;
+				boolean response = false;
+				LOG.debug("========STEP:1==============");
+				//parent loop
+				for (int j = 1; j <= dividedValue; j++)
 				{
 					result1 = null;
 					response = false;
-
-					//for (int i = rowCount - remainder; i <= rowCount; i++)
+					//for (int i = 1; i <= b; i++)
 					//{
-					result1 = result.subList(remStart, remEnd);
+					result1 = result.subList(a, b);
 					response = sendSmsService.sendBulkSms(result1);
 					//}
 					///delete dynamic query
@@ -125,14 +95,57 @@ public class SendRefundSmsCronJob extends AbstractJobPerformable<CronJobModel>
 							deleteDynamicQuery.append(",");
 						}
 					}
+					a += batch;
+					b += batch;
+					if (b > rowCount)
+					{
+						remStart = (rowCount - a) + 1;
+						remEnd = rowCount;
+						break;
+					}
 				}
+				LOG.debug("========STEP:2==============");
+				//Remainder loop
+				if (remainder != 0)
+				{
+					for (int j = 1; j <= 1; j++)
+					{
+						result1 = null;
+						response = false;
+
+						//for (int i = rowCount - remainder; i <= rowCount; i++)
+						//{
+						result1 = result.subList(remStart, remEnd);
+						response = sendSmsService.sendBulkSms(result1);
+						//}
+						///delete dynamic query
+						if (response)
+						{
+							for (final BulkSmsPerBatch obj : result1)
+							{
+								deleteDynamicQuery.append("'");
+								deleteDynamicQuery.append(obj.getTxnId());
+								deleteDynamicQuery.append("'");
+								deleteDynamicQuery.append(",");
+							}
+						}
+					}
+				}
+				String subQuery = null;
+				if (deleteDynamicQuery.length() != 0)
+				{
+					subQuery = deleteDynamicQuery.substring(0, deleteDynamicQuery.length() - 1);
+					///delete rows
+					LOG.debug("======== Delete query==========" + subQuery);
+					refundSmsDao.deleteRows(subQuery);
+				}
+				LOG.debug("Finished executing perform method of SendRefundSmsCronJob class...");
+				return new PerformResult(CronJobResult.SUCCESS, CronJobStatus.FINISHED);
 			}
-			final String subQuery = deleteDynamicQuery.substring(0, deleteDynamicQuery.length() - 1);
-			///delete rows
-			LOG.debug("======== Delete query==========" + subQuery);
-			refundSmsDao.deleteRows(subQuery);
-			LOG.debug("Finished executing perform method of SendRefundSmsCronJob class...");
-			return new PerformResult(CronJobResult.SUCCESS, CronJobStatus.FINISHED);
+			else
+			{
+				return new PerformResult(CronJobResult.SUCCESS, CronJobStatus.FINISHED);
+			}
 		}
 		catch (final Exception ex)
 		{
